Request Detail Add/Update

+++ EPMS.Interfaces/Repository/IEmployeeRequestRepository.cs
@@ -2,7 +2,7 @@
 
 namespace EPMS.Interfaces.Repository
 {
-    public interface IEmployeeRequestRepository : IBaseRepository<EmployeeRequest, int>
+    public interface IEmployeeRequestRepository : IBaseRepository<EmployeeRequest, long>
     {
     }
 }
